Refaktorert de forskjellige typene varsler inn i en samlet VarselBoks.tsx fil, med tilhørende css under VarselBoks.module.scss filen
