--- src/komponenter/header/header-regular/common/varsler/varsel-typer/beskjed/Beskjed.scss
@@ -1,119 +1,0 @@
-.beskjed {
-  display: flex;
-  flex-direction: row;
-  background-color: #ffffff;
-  border-radius: 8px;
-  margin-bottom: 4px;
-  text-decoration: none;
-  width: 444px;
-
-  transition: background-color 0.1s linear;
-
-  &__tittel {
-      font-size: 18px;
-      font-weight: bolder;
-      color: #272727;
-      padding: 12px 0 5px 0;
-  }
-
-  &__content-wrapper {
-      width: 100%;
-      margin-right: 16px;
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      align-items: center;
-  }
-
-  &__dato {
-      font-size: 16px;
-      color: #707070;
-      padding-bottom: 12px;
-  }
-
-  &__ikon {
-      margin: 0 1rem 0 0;
-      background-color: #99C4DD;
-      width: 8px;
-      border-radius: 8px 0 0 8px;
-  }
-
-  @media (max-width: 476px) {
-      width: 100%;
-  }
-}
-
-.chevron {
-  color: #707070;
-  width: 21px;
-  min-width: 21px;
-  height: 21px;
-}
-
-.beskjed:hover {
-  box-shadow: 0px 4px 6px -1px rgba(0, 0, 0, 0.1), 0px 2px 4px -2px rgba(0, 0, 0, 0.1);
-  cursor: pointer;
-  background-color: #99C4DD;
-
-  .chevron {
-      transition-duration: 0.2s;
-      transform: translate(3px, 0);
-  }
-}
-
-.arkiverbar-beskjed {
-  display: flex;
-  flex-direction: row;
-  background-color: #ffffff;
-  border-radius: 8px;
-  margin-bottom: 4px;
-  width: 444px;
-
-  &__tittel {
-      font-size: 18px;
-      font-weight: normal;
-      color: #272727;
-      padding-bottom: 5px;
-  }
-
-  &__content-wrapper {
-      width: 100%;
-      margin-right: 16px;
-      display: flex;
-      flex-direction: column;
-      padding: 12px 0px;
-  }
-
-  &__dato-og-knapp {
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      align-items: center;
-  }
-
-  &__dato {
-      font-size: 16px;
-      color: #707070;
-  }
-
-  &__ikon {
-      margin: 0 1rem 0 0;
-      background-color: #99C4DD;
-      width: 8px;
-      border-radius: 8px 0 0 8px;
-  }
-
-  @media (max-width: 476px) {
-      width: 100%;
-  }
-}
-
-.arkiverbar-beskjed-hover {
-  display: flex;
-  flex-direction: row;
-  background-color: #ffffff;
-  border-radius: 8px;
-  margin-bottom: 4px;
-  width: 444px;
-  box-shadow: 0px 4px 6px -1px rgba(0, 0, 0, 0.1), 0px 2px 4px -2px rgba(0, 0, 0, 0.1);
-}
