--- src/komponenter/header/header-regular/common/varsler/varsel-typer/oppgave/Oppgave.scss
@@ -1,62 +1,0 @@
-.oppgave {
-  display: flex;
-  flex-direction: row;
-  background-color: #ffffff;
-  border-radius: 8px;
-  margin-bottom: 4px;
-  width: 444px;
-  text-decoration: none;
-
-  transition: background-color 0.1s linear;
-
-  &__tittel {
-      font-size: 18px;
-      font-weight: bolder;
-      color: #272727;
-      padding: 12px 0 5px 0;
-  }
-
-  &__content-wrapper {
-      width: 100%;
-      margin-right: 16px;
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      align-items: center;
-  }
-
-  &__dato {
-      font-size: 16px;
-      color: #707070;
-      padding-bottom: 12px;
-  }
-
-  &__ikon {
-      margin: 0 1rem 0 0;
-      background-color: #ffd799;
-      width: 8px;
-      border-radius: 8px 0 0 8px;
-  }
-
-  @media (max-width: 476px) {
-      width: 100%;
-  }
-}
-
-.chevron {
-  color: #707070;
-  width: 24px;
-  min-width: 24px;
-  height: 24px;
-}
-
-.oppgave:hover {
-  box-shadow: 0px 4px 6px -1px rgba(0, 0, 0, 0.1), 0px 2px 4px -2px rgba(0, 0, 0, 0.1);
-  cursor: pointer;
-  background-color: #ffd799;
-
-  .chevron {
-      transition-duration: 0.2s;
-      transform: translate(3px, 0);
-  }
-}
